requirements.txt
llama-cpp-python>=0.2.66
typer[all]>=0.12.3
GitPython>=3.1.44
rich>=13.7.0
tqdm>=4.67.1
termcolor>=2.3.0
openai-tokenizer>=0.1.3
pytest>=8.4.1
faiss-cpu>=1.11.0
datasets>=3.6.0
pandas>=2.3.0
langchain>=0.3.26
langchain-core>=0.3.66
